Add `CallHints` structure to specify optimization hints for RPC calls.
For now, we have one boolean: `noPromisePipelining`, which specifies that no pipelined calls will be made.

This commit does not actually implement the new boolean, it only strings the hints through all the layers of the stack. This is noisy but mechanical work. Subsequent commits will actually set this flag and respond to it.